接下来把代码重写

+++ src/libs/string.c
@@ -61,7 +61,7 @@ strlen(const char  *ptr)
 	}
 }
 
-int32_t strnlen(int8_t *dest, int32_t length)
+size_t strnlen(int8_t *dest, int32_t length)
 {
 	
 	dest[length] = '\0';
